+++ plugins/bash/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @matatbread/matbot-tool-bash
 
+## 0.4.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @matatbread/matbot-plugin-api@0.4.2
+
 ## 0.3.10
 
 ### Patch Changes